+++ verik-compiler/src/test/kotlin/io/verik/compiler/interpret/StructUnionInterpreterStageTest.kt
@@ -20,7 +20,7 @@ import io.verik.compiler.test.BaseTest
 import io.verik.compiler.test.findDeclaration
 import org.junit.jupiter.api.Test
 
-internal class StructInterpreterStageTest : BaseTest() {
+internal class StructUnionInterpreterStageTest : BaseTest() {
 
     @Test
     fun `interpret struct`() {
@@ -28,8 +28,19 @@ internal class StructInterpreterStageTest : BaseTest() {
             """
                 class S(var x: Boolean): Struct()
             """.trimIndent(),
-            StructInterpreterStage::class,
+            StructUnionInterpreterStage::class,
             "Struct(S, S, [Property(x, Boolean, null, 1, 0)])"
         ) { it.findDeclaration("S") }
+    }
+
+    @Test
+    fun `interpret union`() {
+        driveElementTest(
+            """
+                class U(var x: Boolean, var y: Ubit<`1`>): Union()
+            """.trimIndent(),
+            StructUnionInterpreterStage::class,
+            "Union(U, U, [Property(x, Boolean, null, 1, 0), Property(y, Ubit<`1`>, null, 1, 0)])"
+        ) { it.findDeclaration("U") }
     }
 }